refactor(planner,sensor): split oversized files into focused modules
Split planner/engine.py (952 lines) into three focused modules:
- planner/slot_population.py (~330 lines)
  Slot generation, price/solcast/consumption population, battery capacity
  simulation, and the spike-aware weighted-average algorithm. All pure.
- planner/charge_scheduler.py (~370 lines)
  Discharge window detection, three-priority charge scheduling, excess
  export logic, and seasonal optimization strategy. All pure.
- planner/engine.py (~270 lines)
  Thin orchestrator: calls pipeline stages in order, derives windows,
  returns PlannerOutput. No business logic of its own.

Split custom_sensors/state_collector.py (702 lines) into two modules:
- custom_sensors/config_reader.py (~320 lines)
  build_sensor_config() + build_battery_schedules(). Pure sync functions,
  no HA entity I/O. Fully testable without a running HA instance.
- custom_sensors/state_collector.py (~320 lines)
  async_collect_live_state() + private helpers. Re-exports
  build_sensor_config and build_battery_schedules for backward compat.

Removed duplication:
- hourly_data_populator._compute_weighted_average() was a verbatim copy
  of the planner algorithm. Deleted; now delegates to
  planner.slot_population.weighted_avg_consumption (single source of truth).
- Updated test to import weighted_avg_consumption from planner.slot_population.

Final file sizes (all under 500 lines):
  engine.py: 273 | slot_population.py: 332 | charge_scheduler.py: 373
  config_reader.py: 319 | state_collector.py: 322

249 tests pass. ruff clean.
